chore(cluster): use domain-middleware properly, so domains are used to restart cluster workers
