+++ package.json
@@ -40,7 +40,7 @@
     "pretty-data": "git://github.com/vkiryukhin/pretty-data#master",
     "js2xmlparser": "~0.1.2",
     "mongoose": "~3.8.1",
-    "domain-middleware": "~0.1.0",
+    "domain-middleware": "git://github.com/HabitRPG/domain-middleware#master",
     "swagger-node-express": "git://github.com/lefnire/swagger-node-express#habitrpg",
     "passport": "~0.1.18",
     "passport-facebook": "~1.0.2",
