+++ src/utils.js
@@ -2,7 +2,6 @@ var nodemailer = require('nodemailer');
 var nconf = require('nconf');
 var crypto = require('crypto');
 var path = require("path");
-var cluster = require("cluster");
 
 module.exports.sendEmail = function(mailData) {
   var smtpTransport = nodemailer.createTransport("SMTP",{
@@ -47,38 +46,3 @@ module.exports.setupConfig = function(){
   if (nconf.get('NODE_ENV') === 'production')
     require('newrelic');
 };
-
-module.exports.crashWorker = function(server,mongoose) {
-  return function(err, req, res, next) {
-    if (!cluster.isMaster) {
-      // make sure we close down within 30 seconds
-      var killtimer = setTimeout(function() {
-          process.exit(1);
-      }, 30000);
-      // But don't keep the process open just for that!
-      killtimer.unref();
-      // stop taking new requests.
-      server.close();
-      mongoose.connection.close();
-      cluster.worker.disconnect();
-    }
-    next(err);
-  };
-}
-
-
-module.exports.errorHandler = function(err, req, res, next) {
-  // when we hit an error, send it to admin as an email. If no ADMIN_EMAIL is present, just send it to yourself (SMTP_USER)
-  var stack = (err.stack ? err.stack : err.message ? err.message : err) +
-    "\n ----------------------------\n" +
-    "\n\noriginalUrl: " + req.originalUrl +
-    "\n\nauth: " + req.headers['x-api-user'] + ' | ' + req.headers['x-api-key'] +
-    "\n\nheaders: " + JSON.stringify(req.headers) +
-    "\n\nbody: " + JSON.stringify(req.body) +
-    (res.locals.ops ? "\n\ncompleted ops: " + JSON.stringify(res.locals.ops) : "");
-  var logging = require('./logging');
-  logging.error(stack);
-  var message = err.message ? err.message : err;
-  message =  (message.length < 200) ? message : message.substring(0,100) + message.substring(message.length-100,message.length);
-  res.json(500,{err:message}); //res.end(err.message);
-}
